// src/app/views/manufacturing-planning-system/advance-manufacturing/models/similation-entity.model.ts
export interface SimulationEntity {
  active?: boolean;
  createDate?: string;
  employeeId?: number;
  orderByDirection?: string;
  orderByProperty?: string;
  organizationId?: number;
  pageNumber?: number;
  pageSize?: number;
  plantId?: number;
  query?: string;
  scheduleSimulationId?: number;
  scheduleSimulationName?: string;
  scheduleSimulationNo?: string;
  updateDate?: string;
  note?: string;
}